widgets: node_graph: refactor: change how the elements are drawn
The UI is done in a 2 pass approach. In the first pass the layering of
the elements is setup along with elements that have to be drawn in the
first pass only (such as the user defined parameter ui contents). In
the second pass, the elements are actually drawn, since the
positioning of the elements is now fixed after the first pass. This is
possible since an Shape::Noop can be added instead of the actual shape
to establish the layering and this shape can then be swapped out for
what is actually required.

The previous approach made it impossible to allow the user to provide
custom UI contents for the parameters and was super restrictive.

There is one downside to this approach, zooming the node graph in or
out cannot be support since the elements provided by the user cannot
be scaled. This is unfortunate but a necessary change to allow much
more customization. This downside could go away in the future when
egui supports scaling of the individual elements.