widgets: node_graph: make node related changes
Make node position lie on left edge - It is not possible to always
make the node position be the top left corner of the node that is
drawn but it ensures that the node position is always on the left edge
of the node.

If possible, force width of the background rect of the node be the
width of the node - Due to the above (node position change), it is
possible to determine the rough size and ensure that the parameter
shapes (along with it's padding) is considered and forced to be within
the bounds of the node width provided, the parameter contents don't
force the node to expand. The next part solves this.

Set node width before interaction - Before interacting with anything,
the node width is set to the drawn background width (due to the above
changes, it is guaranteed that the background width will be exactly
the node width as long as the contents of the parameters don't exceed
the specified size.